Create and open
Added second constructor to Simulation which takes in a filename, to begin the simulation of that map.

Added a second go to simulation method in app which also passes through it the file name.

In start menu boolean determines if we also want to run the simulation.

Draft atm as I haven't tested if this opens the and runs the simulation of the map correctly and maybe the implementation could be changed to be nicer!