made GameEngine not set direction for Aco-agent
Was not able to rework AcoAgnet to use the Move-class as a move, as it used it to store it's previous position. And moving that previous position to another variable somehow also didn't work. And don't want to mess with it too much, as it's not something I was really involved with.